src/modules/finish_withdraw/app/finish_withdraw_presenter.py
from .finish_withdraw_controller import FinishWithdrawController
from .finish_withdraw_usecase import FinishWithdrawUsecase
from src.shared.environments import Environments
from src.shared.helpers.external_interfaces.http_lambda_requests import LambdaHttpRequest, LambdaHttpResponse

repo = Environments.get_withdraw_repo()()
usecase = FinishWithdrawUsecase(repo=repo)
controller = FinishWithdrawController(usecase=usecase)

def lambda_handler(event, context):
    httpRequest = LambdaHttpRequest(data=event)
    httpRequest.data['requester_user'] = event.get('requestContext', {}).get('authorizer', {}).get('claims', None)
    response = controller(request=httpRequest)
    httpResponse = LambdaHttpResponse(status_code=response.status_code, body=response.body, headers=response.headers)

    return httpResponse.toDict()