tests/modules/finish_withdraw/app/test_finish_withdraw_presenter.py
import json
from src.modules.finish_withdraw.app.finish_withdraw_presenter import lambda_handler
from datetime import datetime

class Test_FinishWithdrawPresenter:
    def test_finish_withdraw_presenter(self):
        event = {
            "version": "2.0",
            "routeKey": "$default",
            "rawPath": "/my/path",
            "rawQueryString": "parameter1=value1&parameter1=value2&parameter2=value",
            "cookies": [
                "cookie1",
                "cookie2"
            ],
            "headers": {
                "header1": "value1",
                "header2": "value1,value2"
            },
            "queryStringParameters": {
                "parameter1": "1"
            },
            "requestContext": {
                "accountId": "123456789012",
                "apiId": "<urlid>",
                "authentication": None,
                "authorizer": {
                    "claims" : {
                        "sub" : "e2d865b1-e0c3-427e-866a-efa4e72e20f9",
                        "name" : "Arthur Trevisan",
                        "email" : "arthur@maua.br",
                        "custom:role" : "EMPLOYEE",
                        "custom:ra" : None
                    }
                },
                "domainName": "<url-id>.lambda-url.us-west-2.on.aws",
                "domainPrefix": "<url-id>",
                "external_interfaces": {
                    "method": "POST",
                    "path": "/my/path",
                    "protocol": "HTTP/1.1",
                    "sourceIp": "123.123.123.123",
                    "userAgent": "agent"
                },
                "requestId": "id",
                "routeKey": "$default",
                "stage": "$default",
                "time": "12/Mar/2020:19:03:58 +0000",
                "timeEpoch": 1583348638390
            },
            "body": '{"num_serie":"34036"}',
            "pathParameters": None,
            "isBase64Encoded": None,
            "stageVariables": None
        }

        response = lambda_handler(event, None)

        assert response["statusCode"] == 200
        assert json.loads(response["body"])["withdraw"]["finish_time"] != None
        assert json.loads(response["body"])["message"] == 'the withdraw has been finished'